refactor(grpc): Standardize token counts to uint32 across all protos
- Update SGLang and TensorRT-LLM protos to use uint32 for:
  - prompt_tokens, completion_tokens, cached_tokens in stream chunks
  - prompt_tokens, completion_tokens, cached_tokens in complete messages
  - sequence_index (TensorRT-LLM) for consistency with index (SGLang/vLLM)
  - embedding_dim in EmbedComplete (SGLang)
- vLLM proto already uses uint32 for these fields
- Update proto_wrapper.rs to return u32 instead of i32
- Remove unnecessary type casts throughout the codebase
- Update golang bindings to use u32 for token counts

+++ model_gateway/src/routers/grpc/common/response_formatting.rs
@@ -17,8 +17,8 @@ use crate::{protocols::common::Usage, routers::grpc::proto_wrapper::ProtoGenerat
 /// # Returns
 /// Usage object with aggregated token counts
 pub(crate) fn build_usage(responses: &[ProtoGenerateComplete]) -> Usage {
-    let total_prompt_tokens: u32 = responses.iter().map(|r| r.prompt_tokens() as u32).sum();
-    let total_completion_tokens: u32 = responses.iter().map(|r| r.completion_tokens() as u32).sum();
+    let total_prompt_tokens: u32 = responses.iter().map(|r| r.prompt_tokens()).sum();
+    let total_completion_tokens: u32 = responses.iter().map(|r| r.completion_tokens()).sum();
 
     Usage {
         prompt_tokens: total_prompt_tokens,
